Separate dashboard views from controller logic
